[cli] Split setup-coding-agents into focused modules
Extract agent resolution, key/team source, human rendering, and machine
output from the command handler into util/ai-gateway/coding-agents.
Behavior-preserving; the command file is now a slim orchestrator.
